pythonProject/requirements.txt
requests
uvicorn
fastapi
pydantic
discord.py